move Sybase-related JDBC types to the package where they belong
