+++ hibernate-community-dialects/src/main/java/org/hibernate/community/dialect/SybaseLegacyDialect.java
@@ -12,14 +12,14 @@ import org.hibernate.dialect.DatabaseVersion;
 import org.hibernate.dialect.DmlTargetColumnQualifierSupport;
 import org.hibernate.dialect.NationalizationSupport;
 import org.hibernate.dialect.SybaseDriverKind;
-import org.hibernate.dialect.SybaseJtdsJsonAsStringArrayJdbcTypeConstructor;
-import org.hibernate.dialect.SybaseJtdsJsonAsStringJdbcType;
-import org.hibernate.dialect.SybaseJtdsLongNVarcharJdbcType;
-import org.hibernate.dialect.SybaseJtdsNCharJdbcType;
-import org.hibernate.dialect.SybaseJtdsNClobJdbcType;
-import org.hibernate.dialect.SybaseJtdsNVarcharJdbcType;
-import org.hibernate.dialect.SybaseJtdsXmlAsStringArrayJdbcTypeConstructor;
-import org.hibernate.dialect.SybaseJtdsXmlAsStringJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsJsonAsStringArrayJdbcTypeConstructor;
+import org.hibernate.dialect.type.SybaseJtdsJsonAsStringJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsLongNVarcharJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsNCharJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsNClobJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsNVarcharJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsXmlAsStringArrayJdbcTypeConstructor;
+import org.hibernate.dialect.type.SybaseJtdsXmlAsStringJdbcType;
 import org.hibernate.dialect.function.CommonFunctionFactory;
 import org.hibernate.dialect.function.CountFunction;
 import org.hibernate.dialect.function.IntegralTimestampaddFunction;
