+++ hibernate-core/src/main/java/org/hibernate/dialect/SybaseDialect.java
@@ -19,6 +19,14 @@ import org.hibernate.dialect.lock.internal.TransactSQLLockingSupport;
 import org.hibernate.dialect.lock.spi.LockingSupport;
 import org.hibernate.dialect.sql.ast.SybaseSqlAstTranslator;
 import org.hibernate.dialect.sql.ast.SybaseSqmToSqlAstConverter;
+import org.hibernate.dialect.type.SybaseJtdsJsonAsStringArrayJdbcTypeConstructor;
+import org.hibernate.dialect.type.SybaseJtdsJsonAsStringJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsLongNVarcharJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsNCharJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsNClobJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsNVarcharJdbcType;
+import org.hibernate.dialect.type.SybaseJtdsXmlAsStringArrayJdbcTypeConstructor;
+import org.hibernate.dialect.type.SybaseJtdsXmlAsStringJdbcType;
 import org.hibernate.dialect.unique.SkipNullableUniqueDelegate;
 import org.hibernate.dialect.unique.UniqueDelegate;
 import org.hibernate.engine.jdbc.dialect.spi.DialectResolutionInfo;
