+++ hibernate-core/src/main/java/org/hibernate/dialect/type/SybaseJtdsJsonAsStringArrayJdbcType.java
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: Apache-2.0
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.dialect;
+package org.hibernate.dialect.type;
 
 import org.hibernate.type.SqlTypes;
 import org.hibernate.type.descriptor.ValueBinder;
@@ -13,7 +13,7 @@ import org.hibernate.type.descriptor.jdbc.BasicBinder;
 import org.hibernate.type.descriptor.jdbc.BasicExtractor;
 import org.hibernate.type.descriptor.jdbc.JdbcType;
 import org.hibernate.type.descriptor.jdbc.JdbcTypeIndicators;
-import org.hibernate.type.descriptor.jdbc.XmlAsStringArrayJdbcType;
+import org.hibernate.type.descriptor.jdbc.JsonAsStringArrayJdbcType;
 
 import java.nio.charset.StandardCharsets;
 import java.sql.CallableStatement;
@@ -22,17 +22,17 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 /**
- * Specialized type mapping for {@code XML} that binds UTF-16LE bytes,
+ * Specialized type mapping for {@code JSON} that binds UTF-16LE bytes,
  * because the jTDS driver can't handle Unicode characters and doesn't support nationalized methods.
  * For extraction, the {@code getString} method works fine though.
  */
-public class SybaseJtdsXmlAsStringArrayJdbcType extends XmlAsStringArrayJdbcType {
+public class SybaseJtdsJsonAsStringArrayJdbcType extends JsonAsStringArrayJdbcType {
 
-	public SybaseJtdsXmlAsStringArrayJdbcType(JdbcType elementJdbcType) {
+	public SybaseJtdsJsonAsStringArrayJdbcType(JdbcType elementJdbcType) {
 		super( elementJdbcType, SqlTypes.NCLOB );
 	}
 
-	public SybaseJtdsXmlAsStringArrayJdbcType(JdbcType elementJdbcType, int ddlTypeCode) {
+	public SybaseJtdsJsonAsStringArrayJdbcType(JdbcType elementJdbcType, int ddlTypeCode) {
 		super( elementJdbcType, ddlTypeCode );
 	}
 
@@ -48,13 +48,13 @@ public class SybaseJtdsXmlAsStringArrayJdbcType extends XmlAsStringArrayJdbcType
 		// but that requires the correct jdbc type code to be available, which we ensure this way
 		if ( needsLob( indicators ) ) {
 			return indicators.isNationalized()
-					? new SybaseJtdsXmlAsStringArrayJdbcType( getElementJdbcType(), SqlTypes.NCLOB )
-					: new SybaseJtdsXmlAsStringArrayJdbcType( getElementJdbcType(), SqlTypes.CLOB );
+					? new SybaseJtdsJsonAsStringArrayJdbcType( getElementJdbcType(), SqlTypes.NCLOB )
+					: new SybaseJtdsJsonAsStringArrayJdbcType( getElementJdbcType(), SqlTypes.CLOB );
 		}
 		else {
 			return indicators.isNationalized()
-					? new SybaseJtdsXmlAsStringArrayJdbcType( getElementJdbcType(), SqlTypes.LONG32NVARCHAR )
-					: new SybaseJtdsXmlAsStringArrayJdbcType( getElementJdbcType(), SqlTypes.LONG32VARCHAR );
+					? new SybaseJtdsJsonAsStringArrayJdbcType( getElementJdbcType(), SqlTypes.LONG32NVARCHAR )
+					: new SybaseJtdsJsonAsStringArrayJdbcType( getElementJdbcType(), SqlTypes.LONG32VARCHAR );
 		}
 	}
 
@@ -65,14 +65,14 @@ public class SybaseJtdsXmlAsStringArrayJdbcType extends XmlAsStringArrayJdbcType
 		}
 		return new BasicBinder<>( javaType, this ) {
 
-			private SybaseJtdsXmlAsStringArrayJdbcType getXmlAsStringArrayJdbcType() {
-				return (SybaseJtdsXmlAsStringArrayJdbcType) getJdbcType();
+			private SybaseJtdsJsonAsStringArrayJdbcType getJsonAsStringArrayJdbcType() {
+				return (SybaseJtdsJsonAsStringArrayJdbcType) getJdbcType();
 			}
 
 			@Override
 			protected void doBind(PreparedStatement st, X value, int index, WrapperOptions options)
 					throws SQLException {
-				final SybaseJtdsXmlAsStringArrayJdbcType jdbcType = getXmlAsStringArrayJdbcType();
+				final SybaseJtdsJsonAsStringArrayJdbcType jdbcType = getJsonAsStringArrayJdbcType();
 				final String xml = jdbcType.toString( value, getJavaType(), options );
 				st.setBytes( index, xml.getBytes( StandardCharsets.UTF_16LE ) );
 			}
@@ -80,7 +80,7 @@ public class SybaseJtdsXmlAsStringArrayJdbcType extends XmlAsStringArrayJdbcType
 			@Override
 			protected void doBind(CallableStatement st, X value, String name, WrapperOptions options)
 					throws SQLException {
-				final SybaseJtdsXmlAsStringArrayJdbcType jdbcType = getXmlAsStringArrayJdbcType();
+				final SybaseJtdsJsonAsStringArrayJdbcType jdbcType = getJsonAsStringArrayJdbcType();
 				final String xml = jdbcType.toString( value, getJavaType(), options );
 				st.setBytes( name, xml.getBytes( StandardCharsets.UTF_16LE ) );
 			}
@@ -105,28 +105,28 @@ public class SybaseJtdsXmlAsStringArrayJdbcType extends XmlAsStringArrayJdbcType
 		}
 		return new BasicExtractor<>( javaType, this ) {
 
-			private SybaseJtdsXmlAsStringArrayJdbcType getXmlAsStringArrayJdbcType() {
-				return (SybaseJtdsXmlAsStringArrayJdbcType) getJdbcType();
+			private SybaseJtdsJsonAsStringArrayJdbcType getJsonAsStringArrayJdbcType() {
+				return (SybaseJtdsJsonAsStringArrayJdbcType) getJdbcType();
 			}
 
 			@Override
 			protected X doExtract(ResultSet rs, int paramIndex, WrapperOptions options) throws SQLException {
 				final String value = rs.getString( paramIndex );
-				return getXmlAsStringArrayJdbcType().fromString( value, getJavaType(), options );
+				return getJsonAsStringArrayJdbcType().fromString( value, getJavaType(), options );
 			}
 
 			@Override
 			protected X doExtract(CallableStatement statement, int index, WrapperOptions options)
 					throws SQLException {
 				final String value = statement.getString( index );
-				return getXmlAsStringArrayJdbcType().fromString( value, getJavaType(), options );
+				return getJsonAsStringArrayJdbcType().fromString( value, getJavaType(), options );
 			}
 
 			@Override
 			protected X doExtract(CallableStatement statement, String name, WrapperOptions options)
 					throws SQLException {
 				final String value = statement.getString( name );
-				return getXmlAsStringArrayJdbcType().fromString( value, getJavaType(), options );
+				return getJsonAsStringArrayJdbcType().fromString( value, getJavaType(), options );
 			}
 
 		};
