+++ hibernate-core/src/main/java/org/hibernate/dialect/type/SybaseJtdsLongNVarcharJdbcType.java
@@ -2,8 +2,9 @@
  * SPDX-License-Identifier: Apache-2.0
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.dialect;
+package org.hibernate.dialect.type;
 
+import org.hibernate.type.SqlTypes;
 import org.hibernate.type.descriptor.java.JavaType;
 import org.hibernate.type.descriptor.jdbc.JdbcType;
 import org.hibernate.type.descriptor.jdbc.JdbcTypeIndicators;
@@ -11,21 +12,21 @@ import org.hibernate.type.descriptor.jdbc.spi.JdbcTypeRegistry;
 
 import java.sql.Types;
 
-public class SybaseJtdsNCharJdbcType extends SybaseJtdsNVarcharJdbcType {
+public class SybaseJtdsLongNVarcharJdbcType extends SybaseJtdsNVarcharJdbcType {
 
-	public static final SybaseJtdsNCharJdbcType JTDS_INSTANCE = new SybaseJtdsNCharJdbcType();
+	public static final SybaseJtdsLongNVarcharJdbcType JTDS_INSTANCE = new SybaseJtdsLongNVarcharJdbcType();
 
-	public SybaseJtdsNCharJdbcType() {
+	public SybaseJtdsLongNVarcharJdbcType() {
 	}
 
 	@Override
 	public int getJdbcTypeCode() {
-		return Types.NCHAR;
+		return Types.LONGNVARCHAR;
 	}
 
 	@Override
 	public String toString() {
-		return "SybaseJtdsNCharJdbcType";
+		return "SybaseJtdsLongNVarcharJdbcType";
 	}
 
 	@Override
@@ -41,8 +42,11 @@ public class SybaseJtdsNCharJdbcType extends SybaseJtdsNVarcharJdbcType {
 		if ( indicators.isLob() ) {
 			jdbcTypeCode = indicators.isNationalized() ? Types.NCLOB : Types.CLOB;
 		}
+		else if ( shouldUseMaterializedLob( indicators ) ) {
+			jdbcTypeCode = indicators.isNationalized() ? SqlTypes.MATERIALIZED_NCLOB : SqlTypes.MATERIALIZED_CLOB;
+		}
 		else {
-			jdbcTypeCode = indicators.isNationalized() ? Types.NCHAR : Types.CHAR;
+			jdbcTypeCode = indicators.isNationalized() ? Types.LONGNVARCHAR : Types.LONGVARCHAR;
 		}
 
 		return jdbcTypeRegistry.getDescriptor( indicators.resolveJdbcTypeCode( jdbcTypeCode ) );
